+++ include/effects/matrix/PatternSMRadialFire.h
@@ -26,9 +26,9 @@ class PatternSMRadialFire : public EffectWithId<PatternSMRadialFire>
 
         const auto& rMap = GFXBase::getPolarMap();
 
-        for (uint8_t x = 0; x < MATRIX_WIDTH; x++)
+        for (int x = 0; x < MATRIX_WIDTH; x++)
         {
-            for (uint8_t y = 0; y < MATRIX_HEIGHT; y++)
+            for (int y = 0; y < MATRIX_HEIGHT; y++)
             {
                 uint8_t angle = rMap[x][y].angle;
                 uint8_t radius = rMap[x][y].unscaled_radius; // Use the unscaled radius
